+++ service/eks/errors.go
@@ -8,6 +8,16 @@ import (
 
 const (
 
+	// ErrCodeAccessDeniedException for service response error code
+	// "AccessDeniedException".
+	//
+	// You don't have permissions to perform the requested operation. The user or
+	// role that is making the request must have at least one IAM permissions policy
+	// attached that grants the required permissions. For more information, see
+	// Access Management (https://docs.aws.amazon.com/IAM/latest/UserGuide/access.html)
+	// in the IAM User Guide.
+	ErrCodeAccessDeniedException = "AccessDeniedException"
+
 	// ErrCodeBadRequestException for service response error code
 	// "BadRequestException".
 	//
@@ -87,6 +97,7 @@ const (
 )
 
 var exceptionFromCode = map[string]func(protocol.ResponseMetadata) error{
+	"AccessDeniedException":                newErrorAccessDeniedException,
 	"BadRequestException":                  newErrorBadRequestException,
 	"ClientException":                      newErrorClientException,
 	"InvalidParameterException":            newErrorInvalidParameterException,